Let the Windows job actually reach the seal test
The seal test reads test-vectors/frames.txt with fopen, MSVC deprecates
that in favour of fopen_s, and /WX makes the deprecation an error — so
the Windows job failed at the build step and never ran a test. CNG's
AES-256-GCM has therefore still not been checked against the golden
vectors, which is the one thing that job exists to do for #113.

_CRT_SECURE_NO_WARNINGS on the seal test alone. Nothing else in this
helper reads a file through the C library — the exe uses Win32 — so the
exemption belongs to the one test rather than to the build.

Also casts the sizeof for BCryptSetProperty's ULONG length, and corrects
the workflow comment that still said the Windows job runs the autostart
ladder and nothing else.

Refs #113

+++ helpers/windows/src/seal_aead.cpp
@@ -31,7 +31,8 @@ struct Provider {
         if (BCryptSetProperty(handle, BCRYPT_CHAINING_MODE,
                               reinterpret_cast<PUCHAR>(const_cast<wchar_t *>(
                                   BCRYPT_CHAIN_MODE_GCM)),
-                              sizeof(BCRYPT_CHAIN_MODE_GCM), 0) != STATUS_SUCCESS) {
+                              static_cast<ULONG>(sizeof(BCRYPT_CHAIN_MODE_GCM)),
+                              0) != STATUS_SUCCESS) {
             BCryptCloseAlgorithmProvider(handle, 0);
             handle = nullptr;
         }
